pyproject.toml
[tool.poetry]
name = "azure-openai-sample"
version = "0.1.0"
description = ""
authors = ["ryota-09 <n.ryo950911@gmail.com>"]
readme = "README.md"
packages = [{include = "demo_app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
openai = {version = "0.28", extras = ["datalib"]}
azure-search-documents = "11.4.0b8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
